requirements.txt
Cython==0.29.35
numpy==1.23.5
pandas==1.5.3
imutils==0.5.4        
gunicorn==20.1.0      
python-dateutil==2.8.2
opencv-python  
boto3==1.34.131
azure-storage-blob==12.15.0
python-dotenv==0.15.0
PyMuPDF==1.21.1
XlsxWriter==3.0.9
pdf2image==1.16.0
ultralytics==8.0.94
python-multipart==0.0.6
fastapi==0.96.0
uvicorn==0.22.0
matplotlib==3.7.1
torch==2.2.1
tensorflow==2.12.0
keras==2.12.0
scikit-learn==1.2.1
elasticsearch==7.9.1
azure-core==1.26.3
mysql-connector-python==8.0.33
pika==1.3.2
langchain-openai==0.1.23
langchain-aws==0.1.17